+++ crates/tusker-migration/src/models.rs
@@ -1,66 +1,66 @@
 use std::collections::BTreeMap;
 
-use crate::db::DbMigration;
-use crate::file::MigrationFile;
+use crate::db::AppliedMigration;
+use crate::source::Migration;
 
 pub(crate) enum MigrationStatus<'a> {
-    Ok(&'a MigrationFile, &'a DbMigration),
-    Mismatch(&'a MigrationFile, &'a DbMigration),
-    NotApplied(&'a MigrationFile),
-    FileMissing(&'a DbMigration),
+    Ok(&'a Migration, &'a AppliedMigration),
+    Mismatch(&'a Migration, &'a AppliedMigration),
+    NotApplied(&'a Migration),
+    FileMissing(&'a AppliedMigration),
 }
 
 #[derive(Debug)]
-pub(crate) struct Migration {
+pub(crate) struct MigrationState {
     pub(crate) number: i32,
-    pub(crate) file: Option<MigrationFile>,
-    pub(crate) db: Option<DbMigration>,
+    pub(crate) migration: Option<Migration>,
+    pub(crate) applied: Option<AppliedMigration>,
 }
 
-impl Migration {
+impl MigrationState {
     pub(crate) fn get_status(&self) -> MigrationStatus<'_> {
-        match (&self.file, &self.db) {
-            (Some(file), Some(db)) => {
-                if file.name == db.name && file.hash == db.hash {
-                    MigrationStatus::Ok(file, db)
+        match (&self.migration, &self.applied) {
+            (Some(migration), Some(applied)) => {
+                if migration.name == applied.name && migration.hash == applied.hash {
+                    MigrationStatus::Ok(migration, applied)
                 } else {
-                    MigrationStatus::Mismatch(file, db)
+                    MigrationStatus::Mismatch(migration, applied)
                 }
             }
-            (Some(file), None) => MigrationStatus::NotApplied(file),
-            (None, Some(db)) => MigrationStatus::FileMissing(db),
+            (Some(migration), None) => MigrationStatus::NotApplied(migration),
+            (None, Some(applied)) => MigrationStatus::FileMissing(applied),
             (None, None) => {
-                panic!("Neither 'file' nor 'db' set. This should never happen.");
+                panic!("Neither 'migration' nor 'applied' set. This should never happen.");
             }
         }
     }
 }
 
 pub(crate) fn combine_migrations(
-    migration_files: &[MigrationFile],
-    db_migrations: &[DbMigration],
-) -> Vec<Migration> {
-    let mut map: BTreeMap<i32, Migration> = BTreeMap::new();
-    for migration_file in migration_files {
+    migrations: &[Migration],
+    applied_migrations: &[AppliedMigration],
+) -> Vec<MigrationState> {
+    let mut map: BTreeMap<i32, MigrationState> = BTreeMap::new();
+    for migration in migrations {
         let _ = map.insert(
-            migration_file.number,
-            Migration {
-                number: migration_file.number,
-                file: Some(migration_file.clone()),
-                db: None,
+            migration.number,
+            MigrationState {
+                number: migration.number,
+                migration: Some(migration.clone()),
+                applied: None,
             },
         );
     }
-    for db_migration in db_migrations {
-        if let Some(migration) = map.get_mut(&db_migration.number) {
-            let _ = migration.db.replace(db_migration.clone());
+    for applied in applied_migrations {
+        if let Some(state) = map.get_mut(&applied.number) {
+            let _ = state.applied.replace(applied.clone());
         } else {
             let _ = map.insert(
-                db_migration.number,
-                Migration {
-                    number: db_migration.number,
-                    file: None,
-                    db: Some(db_migration.clone()),
+                applied.number,
+                MigrationState {
+                    number: applied.number,
+                    migration: None,
+                    applied: Some(applied.clone()),
                 },
             );
         }